Extract time handling functions from reminder to time_util
