+++ src/lan.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from math import floor
 
-from reminder import as_helsinki, as_utc, to_utc
+from time_util import as_helsinki, as_utc, to_utc
 
 def register(client):
     return {
