+++ src/reminder.py
@@ -4,7 +4,7 @@ import traceback
 import discord
 import parsedatetime
 from datetime import datetime
-import pytz
+from time_util import to_utc, to_helsinki, as_utc, as_helsinki
 
 import emoji
 import database as db
@@ -78,22 +78,6 @@ async def process_next_reminder(client):
                 SET reminded = true
                 WHERE reminder_id = $1
             """, id)
-
-def as_helsinki(time):
-    """Interpret a time without timezone as Europe/Helsinki without adjusting time"""
-    return pytz.timezone('Europe/Helsinki').localize(time)
-
-def to_helsinki(time):
-    """Translate existing time with timezone to Europe/Helsinki"""
-    return time.astimezone(tz=pytz.timezone('Europe/Helsinki'))
-
-def as_utc(time):
-    """Interpret a time without timezone as UTC without adjusting time"""
-    return pytz.utc.localize(time)
-
-def to_utc(time):
-    """Translate existing time with timezone to UTC"""
-    return time.astimezone(tz=pytz.utc)
 
 def parse_reminder(text):
     cal = parsedatetime.Calendar()
